apps/sheily_light_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.3
requests
httpx==2.31.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
aiofiles==23.2.1
psutil==5.9.8
PyYAML==6.0.1
APScheduler==3.10.4
